Move groupers to better locations
The groupers belong in the `core` package sine they are not strictly referenced
by the compiler. The lane-splitting group is tied into the Run Scanner
infrastructure, so I have moved that there.
